Moving the example code into the Example namespace
