+++ Assets/NewtonVR/Example/NVRExampleSpawner.cs
@@ -1,21 +1,23 @@
 ﻿using UnityEngine;
 using System.Collections;
-using NewtonVR;
 
-public class NVRExampleSpawner : MonoBehaviour
+namespace NewtonVR.Example
 {
-    public NVRButton Button;
-
-    public GameObject ToCopy;
-    public Transform SpawnLocation;
-
-    private void Update()
+    public class NVRExampleSpawner : MonoBehaviour
     {
-        if (Button.ButtonDown)
+        public NVRButton Button;
+
+        public GameObject ToCopy;
+        public Transform SpawnLocation;
+
+        private void Update()
         {
-            GameObject newGo = GameObject.Instantiate(ToCopy);
-            newGo.transform.position = SpawnLocation.position;
-            newGo.transform.localScale = ToCopy.transform.lossyScale;
+            if (Button.ButtonDown)
+            {
+                GameObject newGo = GameObject.Instantiate(ToCopy);
+                newGo.transform.position = SpawnLocation.position;
+                newGo.transform.localScale = ToCopy.transform.lossyScale;
+            }
         }
     }
 }
